kvmmanager/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "kbdr-kvmmanager"
description = "KBDr KVMManager"
readme = "README.md"
requires-python = ">=3.11"
version = "v0.3.5"
dependencies = [
    "aio_pika",
    "google-cloud-compute",
    "google-auth",
    "aiofiles",
    "google-cloud-storage",
    "jsonschema"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["KBDr.*"]  # ["*"] by default
